Breaks general signing tests into single files
