+++ package.json
@@ -10,10 +10,10 @@
     "precommit": "npm run lint:fix && npm run test",
     "prepublish": "",
     "postinstall": "",
+    "test-signing": "ts-mocha --timeout 120000 -R spec test/signing/index.ts --recursive --exit",
     "test": "ts-mocha --timeout 120000 -R spec test/testAll.ts --recursive --exit",
     "test-btc": "ts-mocha --timeout 180000 -R spec test/testBtc.ts --recursive --exit",
     "test-abi": "ts-mocha --timeout 360000 -R spec test/testAbi.ts --recursive --exit",
-    "test-eth-v2": "ts-mocha --timeout 180000 -R spec test/testEthGeneric.ts --recursive --exit",
     "test-eth": "ts-mocha --timeout 180000 -R spec test/testEth.ts --recursive --exit",
     "test-eth-msg": "ts-mocha --timeout 180000 -R spec test/testEthMsg.ts --recursive --exit",
     "test-generic-signing": "ts-mocha --timeout 180000 -R spec test/testGeneric.ts --recursive --exit",
